Disable parallel tool calls + cleanups
