+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "golem-gpt"
-version = "0.1.3"
+version = "0.2.0"
 authors = [
   { name="Alexey Kinev", email="rudy@05bit.com" },
 ]
